requirements.txt
asgiref==3.2.10
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
decorator==4.4.2
Django==2.2.10
django-filter==2.3.0
djangorestframework==3.11.1
Flask==1.1.1
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Markdown==3.2.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
python-telegram-bot==12.5
pytz==2020.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.0
